+++ overlays/fstab/internal/fstab_test.go
@@ -16,7 +16,7 @@ func Test_fstabOverlay(t *testing.T) {
 	defer env.RemoveAll()
 	env.ImportFile("etc/warewulf/nodes.conf", "nodes.conf")
 	env.ImportFile("etc/warewulf/warewulf.conf", "warewulf.conf")
-	assert.NoError(t, config.Get().Read(env.GetPath("etc/warewulf/warewulf.conf")))
+	assert.NoError(t, config.Get().Read(env.GetPath("etc/warewulf/warewulf.conf"), false))
 	env.ImportFile("var/lib/warewulf/overlays/fstab/rootfs/etc/fstab.ww", "../rootfs/etc/fstab.ww")
 
 	tests := []struct {